+++ tweezer.h
@@ -1,9 +1,9 @@
 class Tweezer{
 	public :
-		float A;//amplitude of the tweezer
-		float N;//normalisation coefficient
+		double A;//amplitude of the tweezer
+		double N;//normalisation coefficient
 		double w;//frequency
-		float p;//phase 
+		double p;//phase 
 		Tweezer():A(42),N(0),w(0),p(0){}
 	
 };
